[Chore #167197841] Updated tests for sign in endpoint
