+++ test/test.js
@@ -26,21 +26,25 @@ export default class Test {
     return seeder.trips.insertData;
   }
 
+  static bookings() {
+    return seeder.bookings.insertData;
+  }
+
   static generateToken(id) {
     return token.generate(id);
   }
 }
 
-/*
-require('./users/signup');
+// require('./users/signup');
 require('./users/signin');
+/*
 require('./users/adminSignin');
 require('./buses/createOne');
 require('./buses/getAll');
 require('./trips/createOne');
 require('./trips/getAll');
-*/
 require('./bookings/createOne');
+*/
 
 export {
   expect,
